lab-04/aes_rsa_socket/requirements.txt
pycryptodome
PyQt5